+++ src/modules/[chain]/tx/indexOld.vue
@@ -1,74 +1,104 @@
 <script lang="ts" setup>
-import { useBaseStore, useFormatter } from '@/stores';
-const props = defineProps(['chain']);
+import { useFormatter } from "@/stores";
+import { computed, onMounted, ref, watchEffect } from 'vue';
+import { useQuery } from '@vue/apollo-composable';
+import gql from 'graphql-tag';
+import { reactive } from "vue";
+import { shortenTxHash } from "@/utils";
+import Pagination from "@/components/pagination/Pagination.vue";
+import TransactionTable from "@/components/TransactionTable.vue";
 
-const base = useBaseStore();
+const props = defineProps(['txs', 'chain'])
 const format = useFormatter();
+
+// Transaction in contract
+const transactions = ref();
+const totalCount = ref();
+const pagination = reactive({
+  limit: 10,
+  offset: 0
+})
+
+const query = gql`
+      query GetTransactions($orderBy: [TransactionsOrderBy!], $first: Int!, $offset: Int!) {
+        transactions(orderBy: $orderBy, first: $first, offset:$offset) {
+          results: nodes {
+            id
+            blockNumber
+            gasUsed
+            timestamp
+            sender
+            fee
+            messages {
+              nodes {
+                type
+                subType
+              }
+            }
+          }
+          totalCount
+        }
+      }
+    `;
+
+const variables = computed(() => {
+  return {
+    // filter: {
+    //   messages: {
+    //     some: {
+    //       block: {
+    //         equalTo: "37406263",
+    //       },
+    //     },
+    //   },
+    // },
+    orderBy: "BLOCK_NUMBER_ASC",
+    first: pagination.limit,
+    offset: pagination.offset
+  }
+})
+
+const { result, refetch } = useQuery(query, variables);
+
+watchEffect(() => {
+  if (result.value) {
+    const data = result.value.transactions.results;
+    transactions.value = data.map((item: any) => ({
+      txhash: shortenTxHash(item?.id),
+      result: "Success",
+      message: format.messages(item.messages?.nodes.map((item: any) =>
+        ({ "@type": item.type, typeUrl: item.type })
+      )),
+      height: item.blockNumber,
+      amount: 0,
+      fee: `${Number(item.fee[0].amount) / 1e6} ${item?.fee[0].denom?.toUpperCase()}`,
+      timestamp: format.toLocaleDate(new Date(Number(item.timestamp)))
+    }));
+    totalCount.value = result.value.transactions.totalCount;
+  }
+})
+
+function handlePagination(page: number) {
+  pagination.offset = page * pagination.limit
+  refetch()
+}
 
 </script>
 
 <template>
-  <div class="m-4 md:m-6 border border-base-400 bg-base-100 rounded-2xl">
-    <div class="bg-base-100 overflow-x-auto rounded-2xl px-5">
-      <table class="table w-full table-compact">
-        <thead class="border border-base-200">
-          <tr>
-            <th class="text-white text-xs font-bold">
-              {{ $t('account.messages') }}
-            </th>
-            <th class="text-white text-xs font-bold" style="position: relative; z-index: 2">
-              {{ $t('account.hash') }}
-            </th>
-            <th class="text-white text-xs font-bold">{{ $t('block.fees') }}</th>
-            <th class="text-white text-xs font-bold" style="position: relative; z-index: 2">
-              {{ $t('account.height') }}
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr v-for="(item, index) in base.txsInRecents" :index="index" class="hover:bg-base-300">
-            <td>
-              <span class="bg-[rgba(180,183,187,0.10)] rounded px-2 py-[1px]">
-                {{ format.messages(item.tx.body.messages) }}
-              </span>
-            </td>
-            <td class="truncate text-link" width="50%">
-              <RouterLink :to="`/${props.chain}/tx/${item.hash}`">{{
-                item.hash
-              }}</RouterLink>
-            </td>
-            <td>{{ format.formatTokens(item.tx.authInfo.fee?.amount) }}</td>
-            <td class="text-sm text-link">
-              <RouterLink :to="`/${props.chain}/block/${item.height}`">{{
-                item.height
-              }}</RouterLink>
-            </td>
-          </tr>
-        </tbody>
-      </table>
-      <div class="p-4 rounded-2xl">
-        <div class="alert relative bg-transparent">
-          <div class="alert absolute inset-x-0 inset-y-0 w-full h-full bg-info opacity-10"></div>
-          <div class="text-info flex gap-2">
-            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
-              class="stroke-current flex-shrink-0 w-6 h-6">
-              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"></path>
-            </svg>
-            <span>{{ $t('block.only_tx') }}</span>
-          </div>
-        </div>
-      </div>
+  <div class="m-4 md:m-6 border border-base-400 bg-base-100 rounded-2xl h-full py-5">
+    <div class="bg-base-100 overflow-x-auto rounded-2xl p-5">
+      <TransactionTable :transaction="transactions" :chain="chain" :txTotal="totalCount" :pagination="pagination"
+        :handlePagination="handlePagination" />
     </div>
   </div>
 </template>
-<!-- 
 
-<route>
-    {
-      meta: {
-        i18n: 'transactions',
-        order: 6
-      }
+<!-- <route>
+  {
+    meta: {
+      i18n: 'transactions',
+      order: 6
     }
+  }
 </route> -->
